Make trace accumulative
