+++ util/log.go
@@ -103,37 +103,17 @@ func logLevelForArea(area string) jww.Threshold {
 // LogLevel sets log level for all loggers
 func LogLevel(defaultLevel string, areaLevels map[string]string) {
 	// default level
-	LogThreshold = LogLevelToThreshold(defaultLevel)
+	LogThreshold = logstash.LogLevelToThreshold(defaultLevel)
 
 	// area levels
 	for area, level := range areaLevels {
 		area = strings.ToLower(area)
-		levels[area] = LogLevelToThreshold(level)
+		levels[area] = logstash.LogLevelToThreshold(level)
 	}
 
 	Loggers(func(name string, logger *Logger) {
 		logger.SetStdoutThreshold(logLevelForArea(name))
 	})
-}
-
-// LogLevelToThreshold converts log level string to a jww Threshold
-func LogLevelToThreshold(level string) jww.Threshold {
-	switch strings.ToUpper(level) {
-	case "FATAL":
-		return jww.LevelFatal
-	case "ERROR":
-		return jww.LevelError
-	case "WARN":
-		return jww.LevelWarn
-	case "INFO":
-		return jww.LevelInfo
-	case "DEBUG":
-		return jww.LevelDebug
-	case "TRACE":
-		return jww.LevelTrace
-	default:
-		panic("invalid log level " + level)
-	}
 }
 
 var uiChan chan<- Param
