+++ util/logstash/levels.go
@@ -1,0 +1,27 @@
+package logstash
+
+import (
+	"strings"
+
+	jww "github.com/spf13/jwalterweatherman"
+)
+
+// LogLevelToThreshold converts log level string to a jww Threshold
+func LogLevelToThreshold(level string) jww.Threshold {
+	switch strings.ToUpper(level) {
+	case "FATAL":
+		return jww.LevelFatal
+	case "ERROR":
+		return jww.LevelError
+	case "WARN":
+		return jww.LevelWarn
+	case "INFO":
+		return jww.LevelInfo
+	case "DEBUG":
+		return jww.LevelDebug
+	case "TRACE":
+		return jww.LevelTrace
+	default:
+		return jww.LevelError
+	}
+}
