+++ util/logstash/log_test.go
@@ -3,6 +3,7 @@ package logstash
 import (
 	"testing"
 
+	jww "github.com/spf13/jwalterweatherman"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -20,12 +21,13 @@ func TestLog(t *testing.T) {
 
 	idx := log.data
 
-	assert.Equal(t, []string{s3, s2, s1}, log.All(nil, nil, 0))
-	assert.Equal(t, []string{s3, s2, s1}, log.All([]string{}, []string{}, 0))
+	assert.Equal(t, []string{s3, s2, s1}, log.All(nil, jww.LevelTrace, 0))
+	assert.Equal(t, []string{s3, s2, s1}, log.All([]string{}, jww.LevelTrace, 0))
 
-	assert.Equal(t, []string{s3, s1}, log.All([]string{"test1"}, nil, 0))
-	assert.Equal(t, []string{s3, s1}, log.All(nil, []string{"TRACE"}, 0))
-	assert.Nil(t, log.All(nil, []string{"FATAL"}, 0))
+	assert.Equal(t, []string{s3, s1}, log.All([]string{"test1"}, jww.LevelTrace, 0))
+	assert.Equal(t, []string{s3, s2, s1}, log.All(nil, jww.LevelTrace, 0))
+	assert.Nil(t, log.All(nil, jww.LevelFatal, 0))
+
 	assert.Equal(t, idx, log.data, "data should not be changed after All() call")
 	assert.Equal(t, []string{"test1", "test2"}, log.Areas())
 }
